Ranking Page
There are little issues and implementations problems due to the differences between the designs and what we see.
Also traductions are not completed.
This push was in order to show the demo.
Some issues from previous commit also fixed.
Next commit will merge with fixed smart contract issues and we'll continue refractoring the code